move `Axis` to common
